antigravity/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
requests>=2.31.0
python-dotenv>=1.0.0
